requirements_api.txt
pandas
tensorflow
keras
matplotlib
scikit-learn
opencv-python
pillow
numpy

# API
fastapi         # API framework
pytz            # time zone management
uvicorn         # web server

# Download the model
google.cloud